Fix SQLite database lock on Reader
Fetching posts for the Reader involves an additional query on each row,
which previously ran into our connection limit to the database and caused
it to lock up. This increases the connection limit from 1 to 2, to allow
this.

This is meant to be a quick, safe fix, but there could always be a better
solution.

Fixes #467