Use datetime picker on Meta page and fix API inconsistency
This changes the Post Update API to use the RFC3339 format, to fix inconsistencies between format of data returned and taken as input (fixes #866).

In the course of fixing this, we switch to the more user-friendly `datetime-local` form element on the Post Metadata page. This makes it easier for users to choose a date and conforms to the more consisten API.